renamed files to remove lp reference
