+++ src/features/parcel/service/parcel.service.test.js
@@ -3,8 +3,8 @@ import {
   getParcelActionsWithAvailableArea
 } from './parcel.service.js'
 import { getAvailableAreaDataRequirements } from '~/src/features/available-area/availableAreaDataRequirements.js'
-import { findMaximumAvailableArea } from '~/src/features/available-area/availableArea.lp.js'
-import { formatExplanationSections } from '~/src/features/available-area/explanations.lp.js'
+import { findMaximumAvailableArea } from '~/src/features/available-area/availableArea.js'
+import { formatExplanationSections } from '~/src/features/available-area/explanations.js'
 import {
   actionTransformer,
   plannedActionsTransformer
@@ -13,8 +13,8 @@ import { vi } from 'vitest'
 
 // Mock the dependencies
 vi.mock('~/src/features/available-area/availableAreaDataRequirements.js')
-vi.mock('~/src/features/available-area/availableArea.lp.js')
-vi.mock('~/src/features/available-area/explanations.lp.js')
+vi.mock('~/src/features/available-area/availableArea.js')
+vi.mock('~/src/features/available-area/explanations.js')
 vi.mock('~/src/features/parcel/transformers/parcelActions.transformer.js')
 vi.mock('~/src/features/data-layers/queries/getDataLayer.query.js')
 vi.mock('~/src/features/rules-engine/rulesEngine.js')
